+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - id tokens include `kid` header with the corresponding client id
 
+### Security
+
+- do not not issue an access token in authorization code and hybrid grants if code was issued to an other client
+
 ## [2.1.4] - 2022-06-07
 
 Note that you must run the `boruta.gen.migration` task to keep your database schema up to date while upgrading to this version.